+++ CHANGELOG.md
@@ -6,10 +6,17 @@ Easy way to scrape wattpad books.
 ### Features
 - Get book by url
 
-## [0.0.4] - Added Search Books
+## [0.0.5] - Added Search Books
 
 ### Major Changes
 - Added search books
+
+### Features
+- Search books
+- Get book by url
+
+## [0.0.6] 
+Added Definition for Books, Authors, and Chapters
 
 ### Features
 - Search books